+++ src/PluginManager.php
@@ -14,15 +14,4 @@ use Estasi\PluginManager\Abstracts;
 final class PluginManager extends Abstracts\PluginManager implements Interfaces\PluginManager
 {
     use Traits\PluginManager;
-
-    /**
-     * @inheritDoc
-     */
-    public function getFilter(string $name, iterable $options = null): Interfaces\Filter
-    {
-        /** @var \Estasi\Filter\Interfaces\Filter $filter */
-        $filter = $this->build($name, $options);
-
-        return $filter;
-    }
 }