+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Changelog
 
+## 1.1.0
+### Added
+
+- Filter `Each`
+
+### Changed
+
+- Method `PluginManager::getFilter` moved to trait `Traits\PluginManager`
+
+### FIXED
+
+- `Trim`, `Ltrim` and `Rtrim` character mask
+
+
 ## 1.0.1
 + Deleted interface `Estasi\Filter\Filtration`